admin form for manage skills functions repaired

+++ about_skill_manage.php
@@ -6,9 +6,9 @@ require 'functions.php';
 $db = new PDO($dsn, $user);
 $db->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
 
-deleteSkill($db, $_POST);
-
-redirectIfStuck ($_POST);
+if ($_POST['delete'] != NULL && $_POST['id'] != NULL) {
+    deleteSkill($db, $_POST);
+}
 
 $singleSkill = getSingleSkill($db, $_POST);
 
@@ -28,13 +28,14 @@ $singleSkill = getSingleSkill($db, $_POST);
         <h2>EDIT SKILL</h2>
         <form method="post" action="about_skill_update.php">
             Skills Name:<br>
-            <input type="text" name="skillName" value="<?php echo $singleSkill['skillName']?>"><br>
+            <input type="text" name="skillName" value="<?php echo $singleSkill['skillName']; ?>"><br>
             Image Source:<br>
-            <input type="text" name="imageSource" value="<?php echo $singleSkill['imageSource']?>"><br>
+            <input type="text" name="imageSource" value="<?php echo $singleSkill['imageSource']; ?>"><br>
             Alternative Image Text:<br>
-            <input type="text" name="alternative" value="<?php echo $singleSkill['alternative']?>"><br><br>
+            <input type="text" name="alternative" value="<?php echo $singleSkill['alternative']; ?>"><br><br>
             <input type="submit" value="Save">
-            <input type="hidden" name="id" value="<?php echo $_POST['id']?>">
+            <input type='hidden' name='edit' value='Edit'>
+            <input type="hidden" name="id" value="<?php echo $singleSkill['id']; ?>">
         </form>
     </div>
     <div class="container">
